+++ src/routing/pages/viewer/components/ViewerRecipe.vue
@@ -20,10 +20,11 @@
         </template>
 
         <template #actions>
-            <div class="flex flex-col items-start gap-y-2">
+            <div class="flex flex-col items-start gap-y-2 w-fit">
                 <CoreButton
                     v-if="recipeInCookbook"
                     secondary
+                    class="w-full"
                     route="recipes.show"
                     :route-params="{ recipe: recipeInCookbook?.slug }"
                 >
@@ -31,26 +32,25 @@
                     <span class="ml-2">{{ $t('viewer.recipe.inCookbook') }}</span>
                 </CoreButton>
 
-                <CoreButton v-else secondary @click="importRecipe()">
+                <CoreButton
+                    v-else
+                    secondary
+                    class="w-full"
+                    @click="importRecipe()"
+                >
                     <i-pepicons-plus class="w-5 h-5" aria-hidden="true" />
                     <span class="ml-1">{{ $t('viewer.recipe.import') }}</span>
                 </CoreButton>
 
-                <div v-if="$browser.supportsPrinting">
-                    <CoreButton secondary @click="printRecipe()">
-                        <i-pepicons-file class="w-4 h-4" aria-hidden="true" />
-                        <span class="ml-2">{{ $t('recipes.print') }}</span>
-                    </CoreButton>
-
-                    <CoreButton
-                        clear
-                        :title="$t('viewer.recipe.print_help')"
-                        :aria-label="$t('viewer.recipe.print_help')"
-                        @click="$ui.showMarkdown($t('viewer.recipe.print_info'))"
-                    >
-                        <i-pepicons-question-circle class="w-5 h-5" aria-hidden="true" />
-                    </CoreButton>
-                </div>
+                <CoreButton
+                    v-if="$browser.supportsPrinting"
+                    class="w-full"
+                    secondary
+                    @click="printRecipe()"
+                >
+                    <i-pepicons-file class="w-4 h-4" aria-hidden="true" />
+                    <span class="ml-2">{{ $t('recipes.print') }}</span>
+                </CoreButton>
             </div>
 
             <ViewerRecipeCreator