+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 
+- Printing on mobile devices.
 - Serializing flash routes.
 
 ## [v0.1.1](https://github.com/NoelDeMartin/umai/releases/tag/v0.1.1) - 2023-03-11