+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.5](https://github.com/NoelDeMartin/umai/releases/tag/v0.1.5) - 2023-12-22
+
+### Changed
+
+- [Christmas attire](https://github.com/NoelDeMartin/umai/tree/v0.1.5/src/assets/icons/umai-xmas.svg).
+
 ## [v0.1.4](https://github.com/NoelDeMartin/umai/releases/tag/v0.1.4) - 2023-11-24
 
 ### Fixed